fix bugs

+++ include/Kmeans.h
@@ -7,14 +7,14 @@ using namespace Eigen;
 struct Kmeans {
   ArrayXd *means;
   ArrayXd *vars;
-  int max_iter;
-  int n_clusters;
+  size_t max_iter;
+  size_t n_clusters;
   int random_seed;
 
   /*!
    * \biref Constructor
    */
-  Kmeans(int n_clusters, int max_iter, int random_seed = -1) 
+  Kmeans(size_t n_clusters, size_t max_iter, int random_seed = -1) 
     : n_clusters{n_clusters},
       max_iter{max_iter},
       random_seed{random_seed} {};
